# v2/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Rest API spec .",
    "title": "Pub/Sub Rest API.",
    "version": "2.0.0"
  },
  "host": "localhost:8089",
  "basePath": "/api/ocloudNotifications/v2",
  "paths": {
    "/create/event/": {
      "post": {
        "description": "If publisher is present for the event, then event creation is success and be returned with Accepted (202).",
        "tags": [
          "event"
        ],
        "summary": "Creates a new event.",
        "operationId": "publishEvent",
        "parameters": [
          {
            "description": "event along with publisher id",
            "name": "event",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "202": {
            "$ref": "#/responses/acceptedReq"
          },
          "400": {
            "$ref": "#/responses/badReq"
          }
        }
      }
    },
    "/publishers/": {
      "get": {
        "description": "If publisher creation is success(or if already exists), publisher will be returned with Created (201).",
        "tags": [
          "publishers"
        ],
        "summary": "Get publishers.",
        "operationId": "getPublishers",
        "responses": {
          "200": {
            "$ref": "#/responses/publishers"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        }
      },
      "post": {
        "description": "If publisher creation is success(or if already exists), publisher will be returned with Created (201).",
        "tags": [
          "publishers"
        ],
        "summary": "Creates a new publisher.",
        "operationId": "createPublisher",
        "parameters": [
          {
            "description": "publisher to add to the list of publishers",
            "name": "publisher",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PubSub"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/pubSubResp"
          },
          "400": {
            "$ref": "#/responses/badReq"
          }
        }
      }
    },
    "/subscriptions": {
      "post": {
        "description": "If subscription creation is success(or if already exists), subscription will be returned with Created (201).",
        "tags": [
          "subscription"
        ],
        "summary": "Creates a new subscription.",
        "operationId": "createSubscription",
        "parameters": [
          {
            "description": "subscription to add to the list of subscriptions",
            "name": "subscription",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PubSub"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/pubSubResp"
          },
          "400": {
            "$ref": "#/responses/badReq"
          }
        }
      }
    },
    "/subscriptions/status": {
      "post": {
        "description": "If publisher status ping is success, call  will be returned with status accepted.",
        "tags": [
          "subscriptions"
        ],
        "summary": "Get status of publishing events.",
        "operationId": "pingForSubscribedEventStatus",
        "parameters": [
          {
            "description": "subscription id to check status for",
            "name": "subscriptionid"
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/pubSubResp"
          },
          "400": {
            "$ref": "#/responses/badReq"
          }
        }
      }
    }
  },
  "responses": {
    "acceptedReq": {
      "description": "Accepted"
    },
    "badReq": {
      "description": "Error Bad Request"
    },
    "eventResp": {
      "description": "PubSub request model"
    },
    "notFoundReq": {
      "description": "Error Not Found"
    },
    "pubSubResp": {
      "description": "publisher/subscription data model"
    }
  }
}